+++ scripts/macros/2024/spells/cloudkill.js
@@ -62,7 +62,7 @@ async function move({trigger: {entity: effect, token}}) {
     function getAllowedMoveLocation(casterToken, template, maxSquares) {
         for (let i = maxSquares; i > 0; i--) {
             let movePixels = i * canvas.grid.size;
-            let ray = new Ray(casterToken.center, template.object.center);
+            let ray = new foundry.canvas.geometry.Ray(casterToken.center, template.object.center);
             let newCenter = ray.project((ray.distance + movePixels)/ray.distance);
             let isAllowedLocation = canvas.visibility.testVisibility(newCenter, {object: template.object});
             if (isAllowedLocation) return newCenter;